+++ src/Server/Models/SubscribedClient.cs
@@ -7,7 +7,7 @@ namespace MiniRedis.Models
         public Socket Socket { get; set; }
         public TaskCompletionSource<string> SubscribedTo { get; init; }
         public DateTimeOffset SubscribedAt { get; set; }
-        public int? TimeoutInSeconds { get; init; }
+        public float? TimeoutInSeconds { get; init; }
         public bool IsExpired => TimeoutInSeconds > 0 && 
                                  (DateTimeOffset.UtcNow - SubscribedAt).TotalSeconds > TimeoutInSeconds;
     }
